Consider device key '' when evaluating for 'default' device
On Chrome, if no media device permissions are given, the devices are
exposed with minimal info:
```
[
  {deviceId:'',kind:'audioinput',label:'',groupId:''},
  {deviceId:'',kind:'videoinput',label:'',groupId:''},
  {deviceId:'',kind:'audiooutput',label:'',groupId:''}
]
```

After the media device permissions have been granted, a `devicechange`
event is triggered and the previous device `''` is supposed to be
removed but never is.

This resulted in duplicate audio for users who are presented with
the microphone permissions on their first call as the previous output
device `''` never got properly removed.